:recycle: 조회 기능을 SQL을 통해 효율적으로 조회하도록 변경

+++ java-to-kotlin-application/src/main/kotlin/com/group/libraryapp/controller/book/BookController.kt
@@ -31,7 +31,7 @@ class BookController(
     }
 
     @GetMapping("/book/loan")
-    fun countLoanedBook(): Int {
+    fun countLoanedBook(): Long {
         return bookService.countLoanedBook()
     }
